Lazy load IVAs
