+++ src/app/work-packages/features/upload-work-package-dialog/upload-work-package-dialog.html
@@ -21,7 +21,7 @@
       {{ grant.valid_until | date: friendlyDateFormat }}.
     </p>
 
-    @if (iva === undefined) {
+    @if (iva() === undefined) {
       <p class="flex items-center">
         <span class="text-error"
           ><mat-icon fontIcon="error" class="me-2 mt-1 shrink-0"></mat-icon></span
@@ -31,14 +31,14 @@
         We could not find an independent verification address (IVA) associated with this
         grant. Please contact the helpdesk for assistance with this issue.
       </p>
-    } @else if (iva.state !== 'Verified') {
+    } @else if (iva()?.state !== 'Verified') {
       <p class="flex items-center">
         <span class="text-warning"
           ><mat-icon fontIcon="warning" class="me-2 mt-1 shrink-0"></mat-icon
         ></span>
         <span
           ><span class="text-warning">Unverified IVA:</span>
-          {{ (iva.type | IvaTypePipe).name }}: {{ iva.value }}</span
+          {{ (iva()!.type | IvaTypePipe).name }}: {{ iva()!.value }}</span
         >
       </p>
       <p
@@ -125,8 +125,8 @@
     Close
   </button>
   @if (
-    iva &&
-    iva.state === 'Verified' &&
+    iva() &&
+    iva()!.state === 'Verified' &&
     !token() &&
     !tokenError() &&
     !tokenIsLoading()
