+++ src/app/work-packages/features/upload-work-package-dialog/upload-work-package-dialog.ts
@@ -9,6 +9,7 @@ import { DatePipe } from '@angular/common';
 import {
   ChangeDetectionStrategy,
   Component,
+  computed,
   inject,
   signal,
   viewChild,
@@ -26,8 +27,8 @@ import {
 } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { Iva } from '@app/ivas/models/iva';
 import { IvaTypePipe } from '@app/ivas/pipes/iva-type-pipe';
+import { IvaService } from '@app/ivas/services/iva';
 import { NotificationService } from '@app/shared/services/notification';
 import { ParagraphsComponent } from '@app/shared/ui/paragraphs/paragraphs';
 import { FRIENDLY_DATE_FORMAT } from '@app/shared/utils/date-formats';
@@ -36,10 +37,6 @@ import { GrantWithBoxInfo } from '@app/upload/models/grant';
 import { UploadWorkPackageRequest } from '@app/work-packages/models/work-package';
 import { WorkPackageService } from '@app/work-packages/services/work-package';
 import { PubkeyFieldComponent } from '@app/work-packages/ui/pubkey-input/pubkey-input';
-
-interface UploadGrantWithIva extends GrantWithBoxInfo {
-  iva?: Iva;
-}
 
 /**
  * Dialog for creating upload tokens to upload files via the GHGA connector.
@@ -69,10 +66,18 @@ export class UploadWorkPackageDialogComponent {
   #dialogRef = inject(MatDialogRef<UploadWorkPackageDialogComponent>);
   #notify = inject(NotificationService);
   #wpService = inject(WorkPackageService);
+  #iva = inject(IvaService);
 
-  protected grant = inject<UploadGrantWithIva>(MAT_DIALOG_DATA);
-  protected iva = this.grant.iva;
+  protected grant = inject<GrantWithBoxInfo>(MAT_DIALOG_DATA);
+  protected iva = computed(() => {
+    if (!this.grant.iva_id || this.#iva.userIvas.error()) return undefined;
+    return this.#iva.userIvas.value().find((i) => i.id === this.grant.iva_id);
+  });
   protected readonly friendlyDateFormat = FRIENDLY_DATE_FORMAT;
+
+  constructor() {
+    this.#iva.loadUserIvas();
+  }
 
   protected model = signal({ pubkey: '' });
   protected pubkeyField = viewChild.required(PubkeyFieldComponent);
@@ -117,7 +122,7 @@ export class UploadWorkPackageDialogComponent {
    * Submit form to create and display a new upload token.
    */
   onCreateToken(): void {
-    if (!this.uploadForm().valid() || this.iva?.state !== 'Verified') return;
+    if (!this.uploadForm().valid() || this.iva()?.state !== 'Verified') return;
 
     const workPackage: UploadWorkPackageRequest = {
       type: 'upload',
